requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fritzconnection==1.4.1
idna==2.10
prometheus-client==0.9.0
PyYAML==5.4.1
requests==2.25.1
urllib3==1.26.3
